more db work
