+++ src/main/java/com/dnd/tools/encounterhelper/monster/CustomMonsterSearchRepositoryImpl.java
@@ -1,5 +1,6 @@
 package com.dnd.tools.encounterhelper.monster;
 
+import com.dnd.tools.encounterhelper.monster.model.Monster;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
