+++ src/main/java/com/dnd/tools/encounterhelper/monster/model/Alignment.java
@@ -1,4 +1,4 @@
-package com.dnd.tools.encounterhelper.monster2;
+package com.dnd.tools.encounterhelper.monster.model;
 
 public enum Alignment {
   LAWFUL,
