+++ src/main/java/com/dnd/tools/encounterhelper/monster/model/BookSource.java
@@ -1,0 +1,11 @@
+package com.dnd.tools.encounterhelper.monster.model;
+
+import javax.persistence.Embeddable;
+import lombok.Data;
+
+@Data
+@Embeddable
+public class BookSource {
+  private String book;
+  private int page;
+}
