+++ src/main/java/com/dnd/tools/encounterhelper/monster/model/DamageType.java
@@ -1,4 +1,4 @@
-package com.dnd.tools.encounterhelper.monster2;
+package com.dnd.tools.encounterhelper.monster.model;
 
 public enum DamageType {
   ACID,
