+++ src/main/java/com/dnd/tools/encounterhelper/monster/model/InnateSpellCasting.java
@@ -1,0 +1,5 @@
+package com.dnd.tools.encounterhelper.monster.model;
+
+public class InnateSpellCasting {
+
+}
