+++ src/main/java/com/dnd/tools/encounterhelper/monster/model/Resistance.java
@@ -1,0 +1,24 @@
+package com.dnd.tools.encounterhelper.monster.model;
+
+import javax.persistence.Embeddable;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import lombok.Data;
+
+@Data
+@Embeddable
+public class Resistance {
+  @Enumerated(EnumType.STRING)
+  private DamageType damageType;
+  private String condition;
+
+  public Resistance(DamageType damageType) {
+    this.damageType = damageType;
+    this.condition = null;
+  }
+
+  public Resistance(DamageType damageType, String condition) {
+    this.damageType = damageType;
+    this.condition = condition;
+  }
+}
