+++ src/main/java/com/dnd/tools/encounterhelper/monster/model/SubEntry.java
@@ -1,4 +1,4 @@
-package com.dnd.tools.encounterhelper.monster2;
+package com.dnd.tools.encounterhelper.monster.model;
 
 import lombok.Data;
