--- src/main/java/com/dnd/tools/encounterhelper/monster2/Ability.java
@@ -1,11 +1,0 @@
-package com.dnd.tools.encounterhelper.monster2;
-
-import java.util.List;
-import lombok.Data;
-
-@Data
-public class Ability {
-
-  private String name;
-  private List<Entry> entries;
-}
