--- src/main/java/com/dnd/tools/encounterhelper/monster2/BookSource.java
@@ -1,9 +1,0 @@
-package com.dnd.tools.encounterhelper.monster2;
-
-import lombok.Data;
-
-@Data
-public class BookSource {
-  private String book;
-  private int page;
-}
