--- src/main/java/com/dnd/tools/encounterhelper/monster2/Entry.java
@@ -1,5 +1,0 @@
-package com.dnd.tools.encounterhelper.monster2;
-
-public abstract class Entry {
-
-}
