--- src/main/java/com/dnd/tools/encounterhelper/monster2/Size2.java
@@ -1,10 +1,0 @@
-package com.dnd.tools.encounterhelper.monster2;
-
-public enum Size2 {
-  TINY,
-  SMALL,
-  MEDIUM,
-  LARGE,
-  HUGE,
-  GARGANTUAN
-}
