--- src/main/java/com/dnd/tools/encounterhelper/monster2/Speed.java
@@ -1,15 +1,0 @@
-package com.dnd.tools.encounterhelper.monster2;
-
-import java.util.Optional;
-import lombok.Data;
-
-@Data
-public class Speed {
-  private Optional<Integer> walk;
-  private Optional<Integer> burrow;
-  private Optional<Integer> climb;
-  private Optional<Integer> fly;
-  private Optional<Integer> hover;
-  private Optional<Integer> swim;
-
-}
